chore: centralize `SearchIndex` opening and promote `VisibilityChecker`
This creates a new function named `open_search_index(indexrel:
&PgRelation)` and changes all the code to use it.

It also promotes the `VisibilityChekcer` to its own module.
