+++ pg_search/src/api/operator/text.rs
@@ -19,7 +19,7 @@ use crate::api::operator::{
     anyelement_text_opoid, anyelement_text_procoid, attname_from_var, estimate_selectivity,
     make_search_config_opexpr_node, ReturnedNodePointer,
 };
-use crate::postgres::utils::{locate_bm25_index, relfilenode_from_search_config};
+use crate::postgres::utils::locate_bm25_index;
 use crate::query::SearchQueryInput;
 use crate::schema::SearchConfig;
 use crate::{nodecast, UNKNOWN_SELECTIVITY};
@@ -95,10 +95,9 @@ pub fn text_restrict(
 
             let (heaprelid, query) = make_query_from_var_and_const(info, var, const_);
             let indexrel = locate_bm25_index(heaprelid)?;
-            let search_config = SearchConfig::from((query, indexrel));
-            let relfilenode = relfilenode_from_search_config(&search_config);
+            let search_config = SearchConfig::from((query, &indexrel));
 
-            estimate_selectivity(heaprelid, relfilenode, &search_config)
+            estimate_selectivity(&indexrel, &search_config)
         }
     }
