+++ pg_search/src/postgres/mod.rs
@@ -29,9 +29,11 @@ mod validate;
 
 pub mod customscan;
 pub mod datetime;
+pub mod index;
 pub mod transaction;
 pub mod types;
 pub mod utils;
+pub mod visibility_checker;
 
 #[repr(u16)] // b/c that's what [`pg_sys::StrategyNumber`] is
 pub enum ScanStrategy {
